jquery.morphCrossfade: Remove 'overflow' fiddling
When we cross-fade back-and-forth quickly, the second time we would
read the "old" overflow value while the first animation was still
ongoing and it was set to 'hidden', preventing us from restoring the
real original value.

But it turns out that none of this code is needed, because jQuery
already does this when animating 'height', and it does it correctly.
So just remove it!

This fixes the "Remove" buttons for each upload disappearing after a
failed submit on the "Details" step (e.g. blocked by AbuseFilter).

Change-Id: Ib57c0222f729c8bfcacb610e35e8473791a8f75c