Make ConcurrentQueue tests more reliable
There's a large amount of setTimout usage in both
ConcurrentQueue and its tests.

FF seems to reliably add setTimeout thingies of equal timeout
in the right order[1], but there've been reports of Chrome
not always doing so. It's best not to rely on it, I guess.

There could be a valid reason that the execution of the
action was wrapped inside a setTimeout (to be added to a
later point in the execution queue, in order not to block
anything else). The comment, however, seemed to indicate
it was just a matter of not getting stuck when the action
executed right away (which has been solved by also looking
at the size of the queue before determining the queue has
completed)

As a result of this changed method, it's now possible
to add a new item right after the last one has completed
& the it'll still be processed.

The test was kind of rigged anyway: as the comment for
incrementallyDelayedPromise indicated, there was no strong
guarantee for the execution order of things in the queue,
which was kind of worked around by increasing the timeout
for every new item. Some tests even had a larger increase
once they started breaking. I guess this behavior was also
a side effect of unreliable setTimeout behavior.

Meanwhile, I've also changed some minor test details:
* startExecuting happens after binding to the 'complete'
  event - this should never be the other way around, since
  actions could immediately finish (not the case for these
  tests, though)
* Removed some assertions for queue.completed === false,
  which also would be untrue in the case of immediately
  completing actions (or potentially flaky tests because
  of setTimeout). And it's pointless to check anyway.

1: https://dxr.mozilla.org/firefox/source/dom/src/base/nsGlobalWindow.cpp#8296
2: http://stackoverflow.com/questions/1776239/are-equal-timeouts-executed-in-order-in-javascript

Bug: T150854
Change-Id: I4dc3651ca871b9235417ae967b4f8a4ff552ed2a