Get rid of upload.providedFile
`upload.file` (set in `fill()`) and `upload.providedFile`
(set in `checkFile()`) are the same File object.
If the upload succeeds, both will be set.
Some places seem to use `providedFile`, where others use
`file`.
In fact, when we create the `upload` object, we can already
assign the File object. We already so so right away anyway
(by calling `fill()` right after creating the object.
Let's just settle on 1 variable that is set immediately!

Change-Id: Idb8a2ceadfe36a84727f80c0fa5dd01a5940ed8a