Fix tutorial step logging
I've removed the skip logic from mw.UploadWizard, since it wouldn't
load the tutorial step if it was to be skipped. This in turn caused
the tutorial step not to be logged.
Keeping up with current practices, the tutorial controller is now
in charge of self-skipping.

The skip preference indicated by the checkmark was also influenced
by UW config (if set to skip, the checkmark would also be checked,
even if user preference said not to skip)

Bug: T150016
Change-Id: I03f98ae7e82df2e4445a7ebf4ba9841418995d39