Use mw.Api instead of XMLHttpRequest
This removes some very specific code (e.g. FormData object)
that is not commonly used in MW, and mw.Api supports already.
It makes it easier to work with, the response is now just
a promise like everywhere else. We no longer need to subscribe
to these events and have our own handler to parse the response.

Bug: T94099
Change-Id: I9c0ef7780dd152d973afac5ff5ea631465d8a15a