Refactor deeds
The existing deed code was pretty hard to follow.

OwnWork & ThirdParty were mostly alright, even though some of
the code for ThirdParty was in the generic UploadWizardDeed
(and none of the other places that used UploadWizardDeed used
that)
But it got mighty complex when dealing with custom deeds,
where methods got overridden in places you wouldn't even expect.

Change-Id: I444a61586b03f3f5391394750267476a2e1830b1