Add params to warning messages about present-day public domain uploads
We didn't supply all params to the license messages for
the warning that is displayed in case of a current-day
upload of a public domain file.

This needs a change in makeMessage to allow raw HTML,
instead of escaping it (for the <a> elements), which
should be fine since we'll only allow it for message
object, which are created in code - they're not random
user input.

Bug: T175827
Change-Id: I985ca8e7c79bcf47fa72e7a3affa73cf9da73853