Clear result cache on checkUnique when upload complete
The checkUnique call in DesinationChecker.js makes sure the
title field for a file doesn't already exist using api calls.
Results from those calls are cached. The cache, however, was
not being cleared when the user uploaded a new file without
reloading the page, and so a new file with the same title
was overwriting the original upload.

Bug: T179967
Change-Id: I5f88cb10b87bb8395200bc01de0cea7b0e9716d4