Improve upload error handling
handleSubmitResult was doing too many things:
- it turned some warnings into recoverable errors
- it updates the upload status in some cases
- it made UI changes on success (show success indicator & message)
- it would resolve when in fact it was dealing with a recoverable error

There is an existing flow for dealing with recoverable errors already,
so I changed the method to only convert warnings into errors: they'll now
be dealt with elsewhere, in that existing bit of code (instead of scattered
across different places)
This makes it possible to resolve/reject sensibly, making it possible to
actually know when an upload was *really* successful. Which, in turn, made
it possible to move the status & UI updates elsewhere, where they make more
sense: UI updates shouldn't be buried this deep down, but should be done
wherever the submit is initiated, so that when that flow changes, it's easy
to change the result as well (e.g. when the upload needs to be followed by
other API calls, we probably don't want to mark it as "successful" just yet)

Bug: T188502
Change-Id: I41bc4f5a1bcfcaf3361dde6da740a7a4c46aa82a