Don't repeat selected languages in language selection for other inputs
It makes no sense to have selected "English" (or any other),
then add a new input, and be able to select it again (often
selected as default)
Such duplicates were ignored anyway, but it's unclear which
was saved. There shouldn't even be an option of selecting
the same language more than once.

With this patch, as soon as a language changes (or one gets
removed), the dropdowns are removed & added again with the
new current state of remaining languages to pick from.

Bug: T190807
Change-Id: I9a62ba4c960c27c80ef3d51ebb02058a5e9bfee2