Don't open ULS until click completed
There are 2 changes:
- Stop using 'focus' to detect tabbing into the "dropdown" element
- Bind that event handler only once, outside the ULS method

Clicking the dropdown also fires a focus event. Combined with a recent
change (initialiseUls being called multiple times, every time a
language is selected), this would cause the focus event to be fired
multiple times on click, causing it to open/close/open erratically.

With this change, there will only be the 'click' event, which is
already handled in ULS, and the tab event.

Bug: T193425
Change-Id: I3c210055e03cec0d95a0f9877e4a292986694f0a