Merge most of UploadWizard's modules into one
ResourceLoader modules are more like "bundles" than "modules" as
ECMAScript defines them. They are supposed to group code that is
always loaded together. And we never load modules such as
"uw.controller.Step" individually, only as a dependency of
"ext.uploadWizard".

This reduces the size of ResourceLoader's startup module, because it
no longer has to list the names of ~30 extra modules and their
dependencies. The startup module is loaded on *all* pages before
*everything* else, so even small reduction in size is a major benefit.

(It also reduces the size of our own code, because we were loading
some of the messages multiple times.)

Also minor tweaks:
* Remove "group": "ext.uploadWizard" from an unrelated style-only module
* Narrow down dependencies on OOUI (we don't use its toolbars)

On my test wiki (with a lot of other extensions), the improvements are:
* Startup module: 16.7 kB → 16.2 kB
* Total size of Special:UploadWizard: 453 kB → 444 kB

Bug: T192623
Change-Id: I076f05fbf52864cf4585c43c31ca5edaedf4b552