Fix positioning of "Remove" button on "Details" step for small images
If the original file is less than 230 pixels wide, the thumbnail shown
on the "Details" step is also smaller than that. For sufficiently
small images the "Remove" button would display next to the thumbnail
rather than underneath.

Also, center the small thumbnails in the available space, rather than
align to the left. This has no effect on thumbnails of files over 230
pixels wide.

Bug: T197924
Change-Id: I65753c960d4cda9e3e3133b883747265f75fd367