Fix Flickr uploads
The initial problem, detailed in the linked Phabricator task, was that
the external deed type was, for some reason, trying to pass in the
entire UploadWizard configuration object rather than the particular
configuration for a type of license. I decided to choose "thirdParty"
for the relevant license type, since any file coming from Flickr is not
an original upload, at least as far as I understand it.

While testing my fix, I discovered that MediaWiki core had a breaking
change to the CategoryMultiselectWidget class that would also fail only
in the case of a Flickr upload. Instead of using setValues, we must now
use addItems. I'm sure there was a good reason for that, and I feel
silly for missing the breaking change announcements.

With that, Flickr uploads should now work properly.

Bug: T183075
Change-Id: I54b23a1ffd9ea534b3f4ad3ab49b89cf1dcc2b6a