Make sure only 1 radio, even if another has the same value, is selected
There's an issue with "pd-usgov", which appears in the Flickr &
USGov license groups. When selecting either one, the other will
also be selected.
Not only is this a confusing user experience, but it has the side-
effect of prepending the {{flickrreview}} template as well, since
that radio button is now also selected when having selected pd-usgov
from the USGov group.

Bug: T201800
Change-Id: I2e7b232d6609d28b9789a36b767ef66cefaf22cc