Fix category multiwidget with default entries
`addItems` expects an array of item objects, but we're feeding
it an array of string values, which it doesn't know how to handle.
`setValue` appears to be the better method here.

Note that this is related to CategoriesDetailsWidget changing from
inheriting from TagMultiselectWidget instead of CapsuleMultiselectWidget.
Relevant updates happened in If414804e8b6ea2359befa6b578414df83b4117fc,
which changed this to `.setValue()` (as I'm doing again in this commit)
A later patch, I54b23a1ffd9ea534b3f4ad3ab49b89cf1dcc2b6a, then changed
`.setValue()` to `.addItems()`, so I'm not 100% certain that changing
it back is alright.

Bug: T202760
Change-Id: Iafd79ff6161a99b6446326cf3c75e8553fa9b6a3