Slightly more robust arrow order
The order of the arrows, right now, was defined by the order the step
controllers were constructed in.
This *could* fail if we construct a step that may not be used, or if
we construct them in a different order.

This will change things so that the step constructor no longer
blindly appends its own arrow, but the wizard actually follows the
order in which steps are linked together & add the arrows.

This is going to be more robust, especially when I'll be adding a
conditional extra step in a later patch.

Change-Id: If96137e95f157d410216e37ce0998badb2013705