mw.FlickrChecker: Use {{flickrreview}}
We (the Commons community) do not feel UploadWizard's client-side
license reviewing trustworthy enough if it is open to users that
are not sysop or license reviewer. Given that it is quite complex
for UW to perform a server side Flickr review in a unspoofable,
secure, and trustworthy manner, it may be desireable that UW give
up on the license reviewing matter entirely.

{{flickrreview}} invokes a bot to perform the review, which in
theory should be more trustworthy than a client-side license review,
and has been in practice the de-facto primary method for Flickr
license review since 2007 on Commons. This has the added benefit
that it is often easier for a community to write a bot to perform
a review, with human review as a fallback, then to implement
server-side review within UW, allowing possibly addition of more
sources to UW at a reduced development cost.

The information on what the license UploadWizard *thinks* (on the
client side) what the license is is not lost with this patch, as
the license template tag immediately precedes {{flickrreview}}.

In theory there may be third-party wikis that uses UW's Flickr
upload that will be affected by this change, but the audience should
be small enough that it should not be complicated for them to move
on to a new template (or even make it a blank template if they
decide to). Besides, all these templates are derived from Commons'
name scheme anyways; I'd imagine they have a similar process with
Flickr imports.

Two alternatives to this solution exist:
* Let a bot process {{FlickrVerifiedByUploadWizard}}: This is
  undesirable because the template literally says 'verified',
  implying trustworthiness; it would be counter-intuitive to
  re-review the file when it has already been reviewed. Besides,
  there are currently tens of thousands of files tagged with this
  and it would cause the bot at least a day to process these files.
* Override mw.FlickrChecker.licenseMaps locally on Commons: This is
  undesirable because for the most of the community having a patch
  locally applied is much easier than having it applied upstream,
  so if any future change is necessary in Commons's licenseMaps the
  local version on Commons and upstream UW's version is highly
  likely to diverge, and in practice cause upstream licenseMaps to
  be less maintained, considering the Commons community was the unit
  who reported UW's missing 'Public Domain Mark' in the map.

Bug: T89131
Bug: T100062
Change-Id: Ide88cf6d9b2c424b8bcefa395691dfd99865fb76
Signed-off-by: YiFei Zhu <zhuyifei1999@gmail.com>