Less agressive rounding of arrow steps width
4 or 5 steps were not an issue, they resulted in 25 or 25% widths,
adding up to a total of 100%.
Now that we're adding a sixth step, rounding this agressively gets
us to an individual step width of 16% and a total of 96%, making
steps bar look significantly smaller than it's supposed to be.

Percentages can have decimal values, so might as well add a little
more precision in there...

Given some spare time, we could refactor this to use flexbox.
I'm happy with this least invasive, minor correction for now.

Bug: T219947
Change-Id: I943c4559bf1b86a4801e1398f6f81dd1ddfee13e