Show uploads in the same order as in other steps
Because we also need to fetch the entity ids for the newly
submitted files, there's no guarantee on the order in which
the statements widgets are created. So:
- we need to make sure they are added in the correct order,
  by index
- we shouldn't add things until the first one is ready: the
  first one to be added will automatically be selected, and
  we want that one to be the actual first upload
- since we could be waiting a little while until the first
  item has been added, let's display a spinner to indicate
  something is happening

And while I shifted some things around, might as well
re-use the array of promises in the onSubmit method, instead
of keeping track of the statements widgets in yet another var.

Bug: T221077
Change-Id: I4a2f0e4fb965fdaac805fda6bff021d3383b863c