Indicate that statements have been submitted on Metadata step
Adds some new methods to the Metadata controller: setPending and
removePending. These can be called from the onSubmit handler method in
the same file.

setPending uses the existing $.createSpinner jQuery plugin to display a
spinner element and a translucent overlay, which covers the entire UI
for this step.

removePending removes this element. This method is passed as an argument to
deferred.always() to ensure this happens regardless of success or failure.

Some style rules have been added as well; flexbox is used to center the
spinner vertically and horizontally in the overlay div.

Bug: T222361
Change-Id: Ic1aa9a94eef0058373f37ebbd4a3360fde35c322