Use parseDom() instead of parse() wherever possible
Code like this:

    $( '<p>' ).append( mw.message( '...' ).parse() )

First generates the DOM tree for the message, then converts it to HTML
code (in parse()), then parses the HTML back to a DOM tree (in
append()), and only then actually appends the new nodes.

By using parseDom(), we skip the middle part, and directly take the
generated DOM tree and append those nodes. This probably has no
noticeable performance impact, but it's a good practice.

Additionally, where plain text is passed in, change some append() calls
to text(), to correctly HTML-escape it.

Change-Id: I5be74659156a6f758589f36abb0cd0e37c685d75