Call StatementWidget.submit with correct baseRevId argument
This binds without argument, so the previous promise's
return value will be the first argument that is passed
to StatementWidget.submit. That one will accept an
(optional) argument, but will expect it to be a number,
the base revision id. We don't want to have that passed
in here, so let's just explicitly bind that method call
with an undefined first argument - then it'll know how
to take it from there (instead of choking on the invalid
argument like it does now)

Change-Id: Ifa82d76e2e15af6608607a215f4b058d76c1f705