Restrict MachineVision CTA based on MachineVision config and user groups
Updates the MachineVision CTA so that it is only shown if configured to
do so in MachineVision extension configuration and if the user is
authorized to use the Computer-Aided Tagging feature.

The necessary extension configuration variables will be exported to JS
in a separate patch to the MachineVision extension. In the meantime, the
CTA will default to not being shown.

Change-Id: Ic31e8c3cd519457af787f3f7e211c70ce2177e14