Fix copy-statement functionality
createStatementWidget expects an argument for the data type,
but it was not receiving that for the copy-statement functionality.

Meanwhile also combined the requires & config values until after
the needed modules (where the values might be attached to - e.g.
wbDataTypes won't be pulled in until wikibase.mediainfo.statements
is) are loaded.

Bug: T232499
Depends-On: I0de6339769275cda4baca168f41ba1eb32c11a79
Change-Id: Ifbdb7317455005ba28b2ddbbf9c3d40b19c80a24