Moving stray license definition from CommonSettings.php
Generally, we keep all the license definitions in
UploadWizard.config.php (even though they are more or less specific
to Commons). This makes things easier to track down and also
prevents accidently deleting i18n messages that are only called
from the configs.

Also, changing the key of the license to match the current template
being used for it. When we remove the old license definition from
CommonSettings.php, we'll also need to update the key called from
licenseGroups. (Don't worry, I'll handle all this after the current
pacth is merged.)

Change-Id: I4ac3602146d90aa7e0599d6d80fd43a24352901a