Make the default upload comment a message
This patch introduces two new messages (for own work and third-party
uploads) that specify the upload comment that should be saved in the
upload log. These messages are localized server-side in order to use
the wiki's language, not the user's language. They are then passed
to the JS code where variable substitution is done. This is also why
these two messages don't allow any {{…}} syntax, as it would break
in really horrible ways.

This is a bit ugly, but I don't have a better idea on how to do it.

Bug: T262252
Change-Id: Ib2f91bedaab83454b52b24a512ae09183fcac7b7