uw.TitleDetailsWidget.js: Throw error-title-invalid for titles including octothorpe
If an octothorpe (# character) appears in a file name, mw.Title assumes
that it creates a URL fragment and helpfully seperates everything after
the # into the Title.fragment property, including the file extension.
This means that code referencing the file extension will throw an
exception, and won't generate a user-visible error message. When there
is no # in the input, mw.Title.fragment is null.

If mw.Title.fragment != null, there must have been a # in the input file
name. # characters are not valid in file names, so makeTitleInFileNS
should return null and mwe-upwiz-error-title-invalid should be thrown.

Bug: T238359
Change-Id: I99e2e499c870ef3b7a9a3a5b916c661f964f1fba