Stop using the legacy ActorMigration field "img_user"
It's traditionally impossible to upload an image as an anon, and prior
to the actor migration, this was enforced by the DB schema which only
linked to the user table. So counting the unique img_actor values is a
fine way to count the number of contributors for display. There's no
need for a join.

Bug: T278917
Change-Id: I9019055ddb8bf39a00203ab4a1092d0fffd97b5f