transports: Fix internal "abort" on status checks during chunked upload
Simple uploads of files within the size threshold are handled directly
by post() which assigns the cancellable XHR to `this.request` just
fine.

But if upload() decides to use chunkedUpload() instead, then its
status checks were assigning a plain Promise to `this.request`,
which is not cancellable. Fix this by assigning the internal request
instead, and still returning the same chained Promise as before to
the caller.

Bug: T262471
Change-Id: I5bed4095bd401293d9160080e8cc65987ba09b1b