Remove eslint warnings on ES2015 code
Object.keys is widely in use and all supported browser support this.
A warning "es-x/no-array-prototype-keys": "warn" is not needed.
Use the default configuration from "wikimedia/client-es5".

Object.values in not in use in this code base.
A warning "es-x/no-array-prototype-values": "warn" is not needed.
Use the default configuration from "wikimedia/client-es5".

Change-Id: Ia5241877931ef5ca691b3ed96f2d1e9ec514f966