Move "files-added" event code so it doesn't trigger twice
This issue exists since patch I051cbb2. Before that patch the "change"
handler was added to a $fileInputCtrl element. That element was
re-created every time the .init() code was called. This is different
since I051cbb2. Since then the handler is still added every time
.init() is called, but the element is not re-created. This results in
the handler being added twice.

Steps to reproduce:
* Go to Special:UploadWizard.
* Select a file that was not uploaded before.
* Click "Continue" once.
* Immediately go "Back".
* Try to add another file. You don't even need to remove the previous
  file. You will now always see the error "You are already uploading
  the file".

Bug: T235542
Change-Id: Id795655c06e30c8d04a4d8e8de6f276581d2dc71