Redesign specifying the license and public domain information for someone else's work
Note: this is a backwards-incompatible change in the
campaigns schema (making 'head' required for license
groups), but there is no existing campaigns schema
where this is not already satisfied.

Note 2: second backwards-incompatible change in the
campaigns schema is changing 'ownWork.default' (string)
to 'ownWork.defaults' (array of string). The former
form was not supported in code, and not used in existing
campaigns. There is 1 campaign using the new form,
though: https://commons.wikimedia.org/wiki/Campaign:Mus%C3%A9e_de_Bretagne

Note 3: third BC-incompatible change in the campaigns
schema: 'thirdParty.defaults' turns into an array of
strings as well (rather than a single string).
Both were (and still are) actively supported in code
(a single string is cast to array), but no campaigns
were using this anyway.

Bug: T347749
Change-Id: Id378fe90b7c003b74be15dacef8e6ad8e6745d20