Make sure links to CC licenses in the upload form are valid
Before this, the upload form would indiscriminately link to
CC license URL + '/deed.' + <user's language code>, regardless
of whether or not that is actually a valid URL on Creative
Commons' websites. All the CC licenses (I checked) are available
in the same set of 47 languages; users using the UploadWizard
in any language besides those 47 would get a 404 if they clicked
one of the "(learn more)" links. This fixes that by using the
language fallback chain, which always has English as the
final fallback, and English is (not surprisingly) guaranteed
to be in the list of translations Creative Commons provides
for its licenses.

Bug: T354225
Change-Id: Icfa5193f81fd85cf97db45cc0da1ad7eb88034f4