Fix automatic numbering of copied titles
Actually setting the title now happens later on,
after the title check has completed. This happens
inside a callback function that used to grab the
"serialized" object literal (which is provided
by `copyMetadata` in uw.CopyMetadataWidget.js,
in a loop.

At the time the callback executes, that object
literal has been changed a couple of times, for
the next couple of titles. As a result, its value
within all of those callbacks will be the same.

Instead of grabbing the value from within the
object literal (which is passed by reference),
we'll extract the actual title value beforehand.

Bug: T365107
Change-Id: I89dcbd9388cc20937d0cb7987f4b89f103706486