Remove complicated synchronization of caption/description inputs
There is a bug with UW's "copy" widget, where it fails
to copy over descriptions: the checkbos (that they are
not the same as captions) is not taken into account, so
copying them is a no-op: they remain "same as caption".

Fixing that would require also serializing the state of
the caption button, and properly handling unserialization
again. All of this could get a little complicated, so I
figured I'd just get rid of the synchronization
complication altogether: we will no longer attempt to
synchronize the caption & description inputs; instead
we'll just ignore descriptions if that "same as captions"
checkbox is checked.

Bug: T365119
Change-Id: Ia4acf5e64bdd0a09cd15bf01685b1e08e83728c7