Remove explicit font-size for fieldset titles
This is mostly a no-op, since the default font size,
at least in both Vector variants, ends up rendering
at 14px.
In other skins (Minerva being one), however, the
default font size ends up being different (larger,
in the case of Minerva), which makes these titles
seem out of place (smaller in the case of Minerva)
compared to the "optional" suffix and the rest of
the text.

Bug: T365662
Change-Id: I003a7236257d912f04a9828a0757958f12081ccc