Remove use of mediawiki.icon
Actually it wasn't used at all, there isn't use of toggleables
triangles in UploadWizard and the places suggest the use are using
OOUI radio buttons, guess there was some historical uses that are
gone for a long time already so this just removes the module's
references.

The use of it should be supposed to be like
I3b7950b301b2d83c8721c4dcab4a5ccdb0246423 in MediaUploader
which is going to be removed also.

Bug: T351681
Change-Id: I172f067623d83036ce5f1e7650a6d6b161c82581