Config: Reset static `$mergedConfig` between tests
Without this, when `$wgUploadWizardConfig` is reset to its default
of empty array, the the one-time merge of UploadWizard.config.php
stored back into that same variable, is lost.

Details on the task. I suggest as future improvement to refactor this
such that the result of loading (or applying) UploadWizard.config.php
is stored within the class, instead of (ab)using the
$wgUploadWizardConfig input as a way to store a derived output value.

Bug: T378299
Change-Id: I263f90c829c3bde8b96c53e7cf75d6795e865e93