Describe step: set max title length to 235 bytes
The max title length depends on the DB column size, which maxes to 240
bytes. However, the file extension isn't included in the title. It's
added later at publication time, so reserve 5 bytes to it (based on
possible extensions as per
https://commons.wikimedia.org/wiki/Special:MediaStatistics).
Also add 1 character to the text input limit, or the relevant error
message will never show up.

Bug: T370259
Change-Id: Icb40f742ae63ba6c2cc88ad6f6a43a7e07e78737