Move errors deeper down
Some errors were being handled on a higher level
than the widget where they occured.
License selection, for example, where an error
in a custom input field is actually 3 levels deep
(radio button, another radio, then the input)
Instead of handling this at the outermost level
(and having the error message be far removed from
the field it occured), the validation checks are
not being moved to the actual element.

In some places, this becomes a little icky, because
it relies on the field being wrapped inside an
uw.Fieldlayout element (which does the error
displaying), which is otherwise unnecessary.

Bug: T381286
Change-Id: Ie6ad0bffc0dc9cce527d852c50beafbed8a440c7