Improve error summary
This also wires up an awful lot of change events
that end up not being used. I initially wanted to
listen for any change event in order to update the
message, but the problem is that validation may
not yet have completed by the time the change event
has propagated (and so the error message is still
visible)

I settled for a MutationObserver in the end, so
all that change event propagation has become
useless. That said, since the work had been done,
I saw no harm in keeping the upward change event
propagation.

Bug: T381333
Change-Id: I84814518e27b72da1065f347f62f09ff7a61ffa1