Fix inconsistent ability to select parent category
This inverts the logic on when to use the clone and
when not to.

Previously, the original object was manipulated and
the clones exclusively used to ensure we pass along
an original copy to the subcategory handler.

However, because the original object is also part of
the 'results' array that we also pass along into the
subcategory handler (for upward navigation), there
was still an edge case (i.e. upward navigation) where
an unwanted manipulation of the data (i.e. adding a
custom handler) still made its way in.

This change simply stops manipulating the original
data altogether and makes changes on a clone. The
clone is fed into the menu option data; the original
data is passed along to the handlers (who then
process it, adding handlers as needed)

Bug: T388152
Change-Id: I7c3264e7a5a729b9deab6fd60d4ee129a30d9776