Clean up outdated code in the Config class
Rewrite outdated and surprisingly overcomplicated code in the basic
Config class.

The definition of what makes an array "associative" here was wrong.
Luckily this is not relevant in real-world scenarios. Config values
that are intentionally constructed to be sparse should never exist.
And even if, the change proposed in this patch makes the behavior
more robust (old values that only exist in the original array are
not removed any more). This behavior also fits the methods's
documentation better.

Change-Id: Id40698a0b2ac9b49b33ee73964d632f2179e506a