Force the add file UI to show if no existing uploads are found
This patch makes it so that whenever the Upload file step is loaded
without any pre-existing uploads, the add file UI is unhidden and shown
to the user. This means that even if the user is on the flickr screen
with no files and presses the previous button followed by the next
button, the add file UI is unhidden and shown instead of a blank space.
The user can then choose to again select the "Share images from Flickr"
button to enter the flick file input state.

Bug: T404933
Change-Id: Id094b9ba10021addfb6f4dc75001eaa8ae2c0903